# postman_automation/api/openapi.yaml
openapi: 3.0.3
info:
  title: Task Manager API
  description: |
    Simple task management API for demonstrating Postman automation.

    ## Authentication Flow
    1. Register a new user → Get user_id
    2. Login with credentials → Get JWT token
    3. Use token in Authorization header for protected endpoints

    ## Typical Usage Flow
    1. POST /api/v1/auth/register
    2. POST /api/v1/auth/login (get token)
    3. POST /api/v1/tasks (with Bearer token)
    4. GET /api/v1/tasks (list all tasks)
    5. PATCH /api/v1/tasks/{id}/status (update status)
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080/api/v1
    description: Local development server

tags:
  - name: Auth
    description: Authentication endpoints
  - name: Tasks
    description: Task management endpoints (requires authentication)
  - name: Health
    description: Service health check

security:
  - bearerAuth: []

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check if the service is running
      operationId: healthCheck
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      description: Create a new user account
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              example1:
                summary: Valid registration
                value:
                  email: user@example.com
                  password: password123
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              examples:
                example1:
                  summary: Successful registration
                  value:
                    user_id: 550e8400-e29b-41d4-a716-446655440000
                    email: user@example.com
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingEmail:
                  summary: Missing email
                  value:
                    error: validation_error
                    message: Email is required
                    code: 400
                missingPassword:
                  summary: Missing password
                  value:
                    error: validation_error
                    message: Password is required
                    code: 400
        '409':
          description: Conflict - user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                userExists:
                  summary: User already exists
                  value:
                    error: user_exists
                    message: User with this email already exists
                    code: 409

  /auth/login:
    post:
      tags:
        - Auth
      summary: Login user
      description: Authenticate and receive JWT token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              example1:
                summary: Valid login
                value:
                  email: user@example.com
                  password: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                example1:
                  summary: Successful login
                  value:
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiZXhwIjoxNzA2NzAwMDAwLCJpYXQiOjE3MDY2MTM2MDB9.example
                    user_id: 550e8400-e29b-41d4-a716-446655440000
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCredentials:
                  summary: Invalid credentials
                  value:
                    error: invalid_credentials
                    message: Invalid email or password
                    code: 401

  /tasks:
    post:
      tags:
        - Tasks
      summary: Create a new task
      description: Create a new task for the authenticated user
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
            examples:
              example1:
                summary: High priority task
                value:
                  title: Implement user authentication
                  description: Add JWT-based authentication to the API
                  priority: high
              example2:
                summary: Simple task
                value:
                  title: Write documentation
                  description: ""
                  priority: medium
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              examples:
                example1:
                  summary: Created task
                  value:
                    id: 123e4567-e89b-12d3-a456-426614174000
                    user_id: 550e8400-e29b-41d4-a716-446655440000
                    title: Implement user authentication
                    description: Add JWT-based authentication to the API
                    status: pending
                    priority: high
                    created_at: "2024-01-31T10:00:00Z"
                    updated_at: "2024-01-31T10:00:00Z"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingTitle:
                  summary: Missing title
                  value:
                    error: validation_error
                    message: Title is required
                    code: 400
                invalidPriority:
                  summary: Invalid priority
                  value:
                    error: validation_error
                    message: Priority must be low, medium, or high
                    code: 400
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Tasks
      summary: Get all tasks
      description: Get all tasks for the authenticated user
      operationId: getTasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
              examples:
                example1:
                  summary: Multiple tasks
                  value:
                    - id: 123e4567-e89b-12d3-a456-426614174000
                      user_id: 550e8400-e29b-41d4-a716-446655440000
                      title: Implement user authentication
                      description: Add JWT-based authentication to the API
                      status: in_progress
                      priority: high
                      created_at: "2024-01-31T10:00:00Z"
                      updated_at: "2024-01-31T11:00:00Z"
                    - id: 987e6543-e21b-12d3-a456-426614174111
                      user_id: 550e8400-e29b-41d4-a716-446655440000
                      title: Write documentation
                      description: ""
                      status: pending
                      priority: medium
                      created_at: "2024-01-31T12:00:00Z"
                      updated_at: "2024-01-31T12:00:00Z"
                emptyList:
                  summary: No tasks
                  value: []
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Get task by ID
      description: Get a specific task by its ID
      operationId: getTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              examples:
                example1:
                  summary: Task found
                  value:
                    id: 123e4567-e89b-12d3-a456-426614174000
                    user_id: 550e8400-e29b-41d4-a716-446655440000
                    title: Implement user authentication
                    description: Add JWT-based authentication to the API
                    status: in_progress
                    priority: high
                    created_at: "2024-01-31T10:00:00Z"
                    updated_at: "2024-01-31T11:00:00Z"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: Task not found
                  value:
                    error: task_not_found
                    message: Task not found
                    code: 404

    put:
      tags:
        - Tasks
      summary: Update task
      description: Update an existing task
      operationId: updateTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
            examples:
              example1:
                summary: Update task details
                value:
                  title: Implement OAuth authentication
                  description: Add OAuth2.0 support in addition to JWT
                  priority: high
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              examples:
                example1:
                  summary: Updated task
                  value:
                    id: 123e4567-e89b-12d3-a456-426614174000
                    user_id: 550e8400-e29b-41d4-a716-446655440000
                    title: Implement OAuth authentication
                    description: Add OAuth2.0 support in addition to JWT
                    status: in_progress
                    priority: high
                    created_at: "2024-01-31T10:00:00Z"
                    updated_at: "2024-01-31T13:00:00Z"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete a task by its ID
      operationId: deleteTask
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /tasks/{id}/status:
    patch:
      tags:
        - Tasks
      summary: Update task status
      description: Update only the status of a task
      operationId: updateTaskStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: string
            format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
            examples:
              toDone:
                summary: Mark as done
                value:
                  status: done
              toInProgress:
                summary: Mark as in progress
                value:
                  status: in_progress
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
              examples:
                example1:
                  summary: Status updated to done
                  value:
                    id: 123e4567-e89b-12d3-a456-426614174000
                    user_id: 550e8400-e29b-41d4-a716-446655440000
                    title: Implement user authentication
                    description: Add JWT-based authentication to the API
                    status: done
                    priority: high
                    created_at: "2024-01-31T10:00:00Z"
                    updated_at: "2024-01-31T14:00:00Z"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidStatus:
                  summary: Invalid status
                  value:
                    error: validation_error
                    message: Status must be pending, in_progress, or done
                    code: 400
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint

  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
          description: User email address (must be unique)
        password:
          type: string
          format: password
          minLength: 6
          example: password123
          description: User password

    RegisterResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Unique user identifier
        email:
          type: string
          format: email
          example: user@example.com

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: password123

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNTUwZTg0MDAtZTI5Yi00MWQ0LWE3MTYtNDQ2NjU1NDQwMDAwIiwiZXhwIjoxNzA2NzAwMDAwLCJpYXQiOjE3MDY2MTM2MDB9.example
          description: JWT authentication token (valid for 24 hours)
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: User identifier

    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          example: Implement user authentication
          description: Task title
        description:
          type: string
          example: Add JWT-based authentication to the API
          description: Detailed task description (optional)
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
          example: high
          description: Task priority level

    UpdateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          example: Implement OAuth authentication
          description: Updated task title
        description:
          type: string
          example: Add OAuth2.0 support in addition to JWT
          description: Updated task description
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
          example: high
          description: Updated task priority

    UpdateStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [pending, in_progress, done]
          example: done
          description: New task status

    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
          description: Unique task identifier
        user_id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
          description: Owner user ID
        title:
          type: string
          example: Implement user authentication
          description: Task title
        description:
          type: string
          example: Add JWT-based authentication to the API
          description: Task description
        status:
          type: string
          enum: [pending, in_progress, done]
          example: in_progress
          description: Current task status
        priority:
          type: string
          enum: [low, medium, high]
          example: high
          description: Task priority level
        created_at:
          type: string
          format: date-time
          example: "2024-01-31T10:00:00Z"
          description: Task creation timestamp
        updated_at:
          type: string
          format: date-time
          example: "2024-01-31T11:00:00Z"
          description: Last update timestamp

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: validation_error
          description: Error code/type
        message:
          type: string
          example: Title is required
          description: Human-readable error message
        code:
          type: integer
          example: 400
          description: HTTP status code